+++ app/models.py
@@ -1,26 +1,16 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
-from sqlalchemy.orm import relationship
-
-from .database import Base
+from pydantic import BaseModel
 
 
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-    is_active = Column(Boolean, default=True)
-
-    items = relationship("Item", back_populates="owner")
+class UserBase(BaseModel):
+    email: str
 
 
-class Item(Base):
-    __tablename__ = "items"
+class UserCreate(UserBase):
+    password: str
 
-    id = Column(Integer, primary_key=True, index=True)
-    title = Column(String, index=True)
-    description = Column(String, index=True)
-    owner_id = Column(Integer, ForeignKey("users.id"))
 
-    owner = relationship("User", back_populates="items")
+class User(UserBase):
+    user_id: int
+
+    class Config:
+        from_attributes = True
